Release/v3.0.0

+++ src/main/java/de/numcodex/feasibility_gui_backend/query/api/QueryResultLine.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @Builder
 public class QueryResultLine {
     private String siteName;
-    private int numberOfPatients;
+    private long numberOfPatients;
 }
